refactor(memory): extract shared MemoryCurationEvaluator from actor
opsx: memory-core-redesign slice 1 (task 1.1)

Extracts the full curation decision flow (record bypass -> fuzzy anchor
candidates -> content-term search -> CurationRulesEvaluator.Evaluate ->
LLM tier + GuardDestructiveUpdate -> TryAutoResolveAmbiguous -> Create
fallback) out of MemoryCurationActor.EvaluateSingleAsync into a new
MemoryCurationEvaluator, and wires the actor through it. The evaluator
also owns ApplyDecisionAsync, the decision -> write-operation mapping
(including Consolidate's re-anchor/tombstone side effects) that used to
live inline in the actor's write phase.

Constructor dependencies are SQLiteMemoryStore (required) and IChatClient?
(genuinely optional: absent LLM is a real permanent operating mode for a
session with no compaction-role model, not a compat shim).

Logging seam: the actor logs via Akka ILoggingAdapter, the daemon engine
(next commit) via Microsoft.Extensions.Logging ILogger. A minimal internal
ICurationLog delegate (AkkaCurationLog / MicrosoftCurationLog) lets one
evaluator body emit identical markers (curation_llm_decision,
curation_llm_no_decision, curation_llm_timeout, curation_llm_error,
curation_dual_search, curation_ambiguous_auto_resolved,
curation_ambiguous_create_fallback, curation_skip/update/consolidate/
create, curation_reanchor, curation_tombstone_anchor) through either
logging stack, since the July 2026 audit tooling greps daemon logs for
these exact marker strings.

SidecarSessionCorrelationTests updated to call
MemoryCurationEvaluator.TryLlmEvaluationAsync (moved from the actor).
